+++ __tests__/controllers/modules/query.test.js
@@ -3,11 +3,11 @@ const {
   paginatedMetaExpression, queryDateExpression, parseQueryDate, parseQuerySort,
   pullPipe_1_matching, pullPipe_2_masking, pullPipe_3_sorting, pullPipe_4_grouping, pullPipe_5_paginating,
   getAggregationQuery,
-} = require(`${__ROOT__}/controllers/interfaces/query`)[Symbol.for('__TEST__')];
+} = require(`${__ROOT__}/controllers/modules/query`)[Symbol.for('__TEST__')];
 
 
 // tests
-describe('Interface: Query', () => {
+describe('Module: Query', () => {
   test('Fn: paginatedMetaExpression', () => {
     // should return an object as $$meta
     // // if all argument are "empty"