+++ controllers/modules/helper.js
@@ -1,0 +1,80 @@
+const { _U_ } = require('../utilities/');
+const { PostsModel } = require('../../models/');
+
+
+/**
+ * Object proxy for case insensitive access
+ */
+const caseInsensitiveProxy = function caseInsensitiveProxy(req, res, next) {
+  req.query = _U_.object.createCaseInsensitiveProxy(req.query);
+  if (typeof next === 'function') return next();
+};
+
+
+/**
+ * HTML title tag modification
+ */
+const modifyHTMLTitleTag = (options) => function modifyHTMLTitleTag(req, res, next) {
+  if (res.locals.$$MODE === 'api') return next();
+
+  const sequence = [];
+  if (options.root !== false) sequence.push(res.locals.$$VIEW.title);
+  if (options.append === true) sequence.push(options.name || options);
+  else sequence.unshift(options.name || options);
+  res.locals.$$VIEW.title = sequence.join(' - ');
+  if (typeof next === 'function') return next();
+};
+
+
+/**
+ * parsed body normalizer
+ */
+const postNormalizer = async (req, res, next) => {
+  if (req.method === 'GET' || !req.body.post) return next();
+  const { post } = req.body;
+
+  // canonicalize                                                                                                       // todo: move to error handler
+  const checkAndResolveConflict = (str) => PostsModel
+    .count({ canonical: { $regex: new RegExp(`^${str}(?:-[0-9]+)?$`) } })
+    .then(counts => {
+      if (counts > 0) return `${str}-${counts + 1}`;
+      return str;
+    });
+
+  // normalize                                                                                                          // todo: only normalize for POST, PATCH (different logic)
+  const normalizedPost = {
+    ...post,
+    // featured: _U_.string.inspectFileURL(post.featured, res.locals.$$SITE.sets.imageTypes),                           // tofix: `inspectFileURL` have been deprecated
+    category: _U_.string.toKebabCase(post.category) || null,
+    content: _U_.string.toEscapedChars(post.content),
+    title: _U_.string.toEscapedChars(post.title),
+    tags: _U_.string.toKebabCase(post.tags) || null,
+  };
+  if (post.state) {
+    normalizedPost.state.forEach(state => {
+      normalizedPost[`state.${state}`] = true;
+    });
+  }
+  if (req.method === 'POST') {
+    normalizedPost.canonical = await checkAndResolveConflict(_U_.string.toKebabCase(post.title));
+  }
+
+  // overwrite
+  req.body.post = normalizedPost;
+
+  return next();
+};
+
+
+// exports
+module.exports = {
+  caseInsensitiveProxy,
+  modifyHTMLTitleTag,
+  postNormalizer,
+};
+
+Object.defineProperty(module.exports, Symbol.for('__TEST__'), {
+  value: {
+    ...module.exports,
+  },
+});