+++ controllers/modules/index.js
@@ -1,6 +1,9 @@
-const regulator = require('./regulator');
-const adapter = require('./adapter');
+const auth = require('./auth');
+const header = require('./header');
+const helper = require('./helper');
+const { paginatedQuery } = require('./query');
+const { parseMultipart } = require('./upload');
 
 
 // exports
-module.exports = { _M_: { ...regulator, ...adapter } };
+module.exports = { _M_: { ...auth, ...helper, ...header, paginatedQuery, parseMultipart } };